# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte Allotjaments Sol",
        "description": "REST API. Projecte Allotjaments Sol. DAW Client i servidor. Per utilitzar les cridades que necesiten token primer\n  s'haura de fer login amb un usuari per obtenir el token i despres utilitzar aquest token a la part on posa Authorize. Per\n  mes informacio, obtenir un usuari o altres peticions utilitzar el link de Contact Equip Sol o mandar un mail a jaumefullana@paucasesnovescifp.cat",
        "contact": {
            "name": "Equip Sol",
            "email": "jaumefullana@paucasesnovescifp.cat"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:81/allotjamentssol/public/"
        }
    ],
    "paths": {
...
        "/api/tarifa": {
            "get": {
                "tags": [
                    "Tarifes"
                ],
                "summary": "Mostrar totes les tarifes.",
                "operationId": "a146d33343c8fe90accc540a78cce2af",
                "responses": {
                    "200": {
                        "description": "Mostrar totes les tarifes."
                    },
                    "400": {
                        "description": "Hi ha un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "Tarifes"
                ],
                "summary": "Modifica una tarifa",
                "description": "Modifica una tarifa.",
                "operationId": "d22f50a499741aa97bfff39abbf6cb0e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "2"
                                    },
                                    "PreuTemporadaAlta": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1000"
                                    },
                                    "PreuTemporadaBaixa": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "800"
                                    },
                                    "IniciTemporadaAlta": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "01/06/23"
                                    },
                                    "FiTemporadaAlta": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "01/10/23"
                                    },
                                    "TipusCategoriesID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Atribut tarifa requerit"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tarifes"
                ],
                "summary": "Crea una tarifa",
                "description": "Crea una nova tarifa.",
                "operationId": "71ce60423ef6f7244de031d5c66f7ff1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "PreuTemporadaAlta": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1000"
                                    },
                                    "PreuTemporadaBaixa": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "800"
                                    },
                                    "IniciTemporadaAlta": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "01/06/23"
                                    },
                                    "FiTemporadaAlta": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "01/10/23"
                                    },
                                    "TipusCategoriesID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Atribut categoria requerit"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tarifes"
                ],
                "summary": "Esborra una tarifa",
                "description": "Esborra una tarifa.",
                "operationId": "91f7cc3f5e800c2fb141c86d0462e58b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Tupla no trobada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tarifa/{id}": {
            "get": {
                "tags": [
                    "Tarifes"
                ],
                "summary": "Mostrar una tarifa",
                "operationId": "57f51d4f2ac0e535983ea32b63c604a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la tarifa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informació de la tarifa.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Hi ha un error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "tarifa no trobada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}